refactor(session): separate grpc adapter and server

effort: max
